clinical_service/requirements.txt
Django>=5.2.1
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
requests>=2.31.0 